+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
 - Naive initial implementation for iOS
+- Allow passing the resolution on start stream